Revert "WD-19720 - Formatting"
This reverts commit 9ae713748d52a52a285b13e64ff58c345d55ec18.
